Move all pages into components
